Move game endpoints the game controller
