+++ routes/game.js
@@ -3,23 +3,68 @@ const express = require('express');
 const router = express.Router();
 
 const gameController = require('../controllers/game')
+const isAuth = require('../middleware/is-auth')
+
+const multer = require('multer');
+
+/********************\\\\\\\\\
+ * Return the page for seeing
+ * all the personal details
+ * of a specific game
+ /*******************/////////
+
+router.get("/details/:gameid", isAuth.isLoggedIn, gameController.details);
+
+/********************\\\\\\\\\
+ * Return the page for editing
+ * all the personal details 
+ * of a specific game
+ /*******************/////////
+
+router.get("/edit/:gameid", isAuth.isLoggedIn, gameController.edit);
+
+/********************\\\\\\\\\
+ * Upload a game by writing
+ * A record into the db
+ /*******************/////////
+
+var storage = multer.diskStorage({ 
+    destination: (req, file, cb) => { 
+        cb(null, 'uploads') 
+    }, 
+    filename: (req, file, cb) => { 
+        cb(null, file.fieldname + '-' + Date.now()) 
+    } ,
+    limits: { fileSize: 50000 }
+}); 
+  
+var upload = multer({ storage: storage }); 
+
+router.post('/Upload', isAuth.isLoggedIn, upload.single('image'), gameController.upload);
 
 /********************\\\\\\\\\
  * Get keybinds for game by name
  /*******************/////////
 
- router.get('/keybinds/:gameName', gameController.keybinds);
+router.get('/keybinds/:gameName', gameController.keybinds);
 
- /********************\\\\\\\\\
-  * Download game by name and version
-  /*******************/////////
- 
-  router.get('/download/:gameName/:versionName?', gameController.download);
+/********************\\\\\\\\\
+ * Download game by name and version
+ /*******************/////////
 
-  /********************\\\\\\\\\
-   * Report game
-   /*******************/////////
-  
-   router.post('/report', gameController.report);
+router.get('/download/:gameName/:versionName?', gameController.download);
+
+/********************\\\\\\\\\
+ * Report game
+ /*******************/////////
+
+router.post('/report', gameController.report);
+
+/********************\\\\\\\\\
+ * Delete a game by deleting
+ * the record in the database
+ /*******************/////////
+
+router.post("/Delete", isAuth.isLoggedIn, gameController.delete)
   
 module.exports = router;
