+++ views/components/addGameForm.ejs
@@ -1,4 +1,4 @@
-<form id="addGameForm" method="POST" action="/user/upload?_csrf=<%= csrfToken %>" enctype="multipart/form-data">
+<form id="addGameForm" method="POST" action="/game/upload?_csrf=<%= csrfToken %>" enctype="multipart/form-data">
     <h6>Game Information</h6>
     <div class="form-group">
         <label for="gameName">Name</label>
