+++ views/components/gameCard.ejs
@@ -1,6 +1,6 @@
 <div id="<%= game.gameInfo.name%>" class="py-2">
     <h3 class="mb-0">
-        <a href='/user/details/<%= game._id %>'>
+        <a href='/game/details/<%= game._id %>'>
             <%= game.gameInfo.name %>
         </a>
     </h3>
