+++ views/components/gameDeleteModal.ejs
@@ -7,7 +7,7 @@
           <span aria-hidden="true">&times;</span>
         </button>
       </div>
-      <form action="/user/delete" method="POST" id="DeleteForm">
+      <form action="/game/delete" method="POST" id="DeleteForm">
         <div class="modal-body">
           Are you sure?
         </div>
