+++ views/includes/nav.ejs
@@ -43,7 +43,7 @@
                 </button>
             <% } %> 
 
-            <a class="btn btn-primary m-2" href="/user/edit/<%= editGameID %>" role="button">Edit</a>
+            <a class="btn btn-primary m-2" href="/game/edit/<%= editGameID %>" role="button">Edit</a>
         <% } %>
         <% if(typeof editUserID != "undefined") { %>
             <a class="btn btn-primary m-2" href="/user/editUser/<%= editUserID %>" role="button">Edit</a>
